Refactor user context code and remove from app.js
