+++ src/App.js
@@ -2,48 +2,24 @@ import styles from "./App.module.css";
 import NavBar from "./components/NavBar";
 import Container from "react-bootstrap/Container";
 import { Route, Switch } from "react-router-dom";
-import axios from "axios";
 
 import "./api/axiosDefaults";
 import SignUpForm from "./pages/auth/SignUpForm";
 import LoginForm from "./pages/auth/LoginForm";
-import { createContext, useState, useEffect } from "react";
-
-export const CurrentUserContext = createContext();
-export const SetCurrentUserContext = createContext();
 
 function App() {
-  const [currentUser, setCurrentUser] = useState(null);
-
-  const handleMount = async () => {
-    try {
-      const { data } = await axios.get("dj-rest-auth/user/");
-      setCurrentUser(data);
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  useEffect(() => {
-    handleMount();
-  }, []);
-
   return (
-    <CurrentUserContext.Provider value={currentUser}>
-      <SetCurrentUserContext.Provider value={setCurrentUser}>
-        <div className={styles.App}>
-          <NavBar />
-          <Container className={styles.Main}>
-            <Switch>
-              <Route exact path="/" render={() => <h1>Home</h1>} />
-              <Route exact path="/login" render={() => <LoginForm />} />
-              <Route exact path="/signup" render={() => <SignUpForm />} />
-              <Route render={() => <p>Page not found.</p>} />
-            </Switch>
-          </Container>
-        </div>
-      </SetCurrentUserContext.Provider>
-    </CurrentUserContext.Provider>
+    <div className={styles.App}>
+      <NavBar />
+      <Container className={styles.Main}>
+        <Switch>
+          <Route exact path="/" render={() => <h1>Home</h1>} />
+          <Route exact path="/login" render={() => <LoginForm />} />
+          <Route exact path="/signup" render={() => <SignUpForm />} />
+          <Route render={() => <p>Page not found.</p>} />
+        </Switch>
+      </Container>
+    </div>
   );
 }
